+++ module/settings.js
@@ -22,8 +22,8 @@ export function registerSettings() {
         type: Number,
     });
     game.settings.register('fallout', "hoversJsonLocation",{
-		name: "JSON file",
-        hint: "Location of the json file with the qualities and effects",
+		name: "Mouse Hover JSON file",
+        hint: "Location of the json file containing the text for qualities and damage effects.",
 		scope: "world",
 		config: true,
 		default: "systems/fallout/assets/hovers.json",		
@@ -32,4 +32,20 @@ export function registerSettings() {
         restricted: true,
         onChange: debounceReload
 	});
+    game.settings.register('fallout', 'gmMomentumShowToPlayers', {
+        name: 'Show Overseer AP To Players',
+        hint: "Shows the Overseer's AP window to everyone. Requires refresh on the players side.",
+        scope: 'world',
+        config: true,
+        default: false,
+        type: Boolean,
+    });
+    game.settings.register('fallout', 'maxAppShowToPlayers', {
+        name: 'Players Can Setup Max AP',
+        hint: "Allows players to settup the Party's MAX AP. Requires refresh on the players side.",
+        scope: 'world',
+        config: true,
+        default: false,
+        type: Boolean,
+    });
 }